Major upgrade of all dependencies
This does a major upgrade of Clash, the libraries, the Haskell package
set, GHC, and Nixpkgs.

GHC is now 8.2.2, and Clash is now at tip-of-trunk. Dependencies on
singletons are dropped, and manual overrides on the package sets are no
longer needed for shake, or th-desugar.

Doctests and Haddocks are fixed for all packages.

clash had to be jailbroken for some spurious bound fixes, but these
should be removed pretty soon.

And the improved nixpkgs snapshot adds some new things too, for example,
the new Z3 4.6.0 release, and newly updated Icestorm tools, etc.

Signed-off-by: Austin Seipp <aseipp@pobox.com>